+++ app/Mail/AccountCreated.php
@@ -21,7 +21,7 @@ class AccountCreated extends Mailable
     public function __construct(Borrower $borrower)
     {
 	    //
-	 $this->borrower = $borrower;
+	    $this->borrower = $borrower;
     }
 
     /**
@@ -32,7 +32,6 @@ class AccountCreated extends Mailable
     public function build()
     {
 	  return $this->view('emails.borrower')
-		  ->text('emails.borrower_plain')
-		  ->with("borrower", $this->borrower);
+		  ->text('emails.borrower_plain');
     }
 }