+++ resources/views/layouts/mainlayout.blade.php
@@ -36,9 +36,6 @@
 	  
 	</div><!-- /.container -->
 
-
-@include('layouts.partials.footer')
-
 @include('layouts.partials.footer-scripts')
 
  </body>